services/authentication/requirements.txt
Flask
PyJWT
Flask-SQLAlchemy
flask_cors
prometheus-flask-exporter
Werkzeug
pymysql
cryptography